Correción de errores en la conexión con la BD.   - Cambiados los tipos de dato float a decimal.   - Agregadas precisiones de columnas deciaml en MEGADbContext.   - Agregado Endpoint para obtener todos los sucriptores en CSuscriptor.
Considerar que, al trabajar con endpoints, se debe especificar qué información se debe
devolver de una consulta (ver endpoint mencionado anteriormente) para evitar errores de
ciclos infinitos.

+++ server/models/Contrato.cs
@@ -6,7 +6,7 @@ public class Contrato()
   public int Idsuscriptor { set; get; }
   public DateTime FechaContr { set; get; }
   public DateTime? FechaFin { set; get; }
-  public int PrecioBase { set; get; }
+  public decimal PrecioBase { set; get; }
 
   // Relaciones.
   public Suscriptor Suscriptor { set; get; } = null!;
