+++ server/models/Promocion.cs
@@ -3,12 +3,12 @@ namespace megaapi.models;
 public class Promocion()
 {
   public int Idpromocion { set; get; }
-  public int Alcance { set; get; }
+  public byte Alcance { set; get; }
   public string Nombre { set; get; } = null!;
   public int? Duracion { set; get; }
   public DateTime FechaRegistro { set; get; }
-  public float PrecioBase { set; get; }
-  public int Tipo { set; get; }
+  public decimal PrecioBase { set; get; }
+  public byte Tipo { set; get; }
   public DateTime Vigencia { set; get; }
 
   // Relaciones.
