+++ server/models/Servicio.cs
@@ -4,7 +4,7 @@ public class Servicio()
 {
   public int Idservicio { set; get; }
   public int? Cantidad { set; get; }
-  public float PrecioBase { set; get; }
-  public int Tipo { set; get; }
+  public decimal PrecioBase { set; get; }
+  public byte Tipo { set; get; }
   public ICollection<PaqueteServicio> Paquetes { set; get; } = null!;
 }
